Added Update3, a more sophisticated simulation technique.
Update3 uses best-fit quadratic to estimate the acceleration
over the interval [0, dt]. Integrating the quadratic gives
a cubic velocity estimate, and integrating that gives a quartic
position estimate. Use this to converge on an updated [pos, vel]
state. This is already 20 times more accurate for calculating
the Earth's orbit than Update2.